+++ Tests/LinuxMain.swift
@@ -468,6 +468,14 @@ extension FileNameRuleTests {
     ]
 }
 
+extension FileTypesOrderRuleTests {
+    static var allTests: [(String, (FileTypesOrderRuleTests) -> () throws -> Void)] = [
+        ("testFileTypesOrderWithDefaultConfiguration", testFileTypesOrderWithDefaultConfiguration),
+        ("testFileTypesOrderReversedOrder", testFileTypesOrderReversedOrder),
+        ("testFileTypesOrderGroupedOrder", testFileTypesOrderGroupedOrder)
+    ]
+}
+
 extension FirstWhereRuleTests {
     static var allTests: [(String, (FirstWhereRuleTests) -> () throws -> Void)] = [
         ("testWithDefaultConfiguration", testWithDefaultConfiguration)
@@ -1281,6 +1289,14 @@ extension TypeBodyLengthRuleTests {
     ]
 }
 
+extension TypeContentsOrderRuleTests {
+    static var allTests: [(String, (TypeContentsOrderRuleTests) -> () throws -> Void)] = [
+        ("testTypeContentsOrderWithDefaultConfiguration", testTypeContentsOrderWithDefaultConfiguration),
+        ("testTypeContentsOrderReversedOrder", testTypeContentsOrderReversedOrder),
+        ("testTypeContentsOrderGroupedOrder", testTypeContentsOrderGroupedOrder)
+    ]
+}
+
 extension TypeNameRuleTests {
     static var allTests: [(String, (TypeNameRuleTests) -> () throws -> Void)] = [
         ("testTypeName", testTypeName),
@@ -1528,6 +1544,7 @@ XCTMain([
     testCase(FileHeaderRuleTests.allTests),
     testCase(FileLengthRuleTests.allTests),
     testCase(FileNameRuleTests.allTests),
+    testCase(FileTypesOrderRuleTests.allTests),
     testCase(FirstWhereRuleTests.allTests),
     testCase(ForWhereRuleTests.allTests),
     testCase(ForceCastRuleTests.allTests),
@@ -1636,6 +1653,7 @@ XCTMain([
     testCase(TrailingSemicolonRuleTests.allTests),
     testCase(TrailingWhitespaceTests.allTests),
     testCase(TypeBodyLengthRuleTests.allTests),
+    testCase(TypeContentsOrderRuleTests.allTests),
     testCase(TypeNameRuleTests.allTests),
     testCase(UnavailableFunctionRuleTests.allTests),
     testCase(UnneededBreakInSwitchRuleTests.allTests),